shell-helper: Use doubles for shell OSD dbus calls
Allow for more precise values to be sent to the shell to fix a bug
where the volume icon in the shells OSD is not the "muted" icon while
the volume bar shows a value of 0.

This happened because casting the volume to an integer value might
return 0 while the value actually is near 0. Since we also generate the
icon name inside media-keys and use the non-casted value for that, the
icon and the volume we show in the shell would be different.

+++ plugins/common/gsd-shell-helper.h
@@ -27,14 +27,14 @@ G_BEGIN_DECLS
 void shell_show_osd (GsdShell    *shell,
 		     const gchar *icon_name,
 		     const gchar *label,
-		     gint         level,
+		     double       level,
 		     gint         monitor);
 
 void shell_show_osd_with_max_level (GsdShell    *shell,
                                     const gchar *icon_name,
                                     const gchar *label,
-                                    gint         level,
-                                    gint         max_level,
+                                    double       level,
+                                    double       max_level,
                                     gint         monitor);
 
 G_END_DECLS
